Cleanup and deleting files
