+++ main.py
@@ -1,9 +1,14 @@
+import sys
+import asyncio
+
+# Fix for Playwright + Streamlit on Windows
+if sys.platform == 'win32':
+    asyncio.set_event_loop_policy(asyncio.WindowsProactorEventLoopPolicy())
+
 import pandas as pd
 import streamlit as st
-import asyncio
 import time
-import asyncio
-from playwright.async_api import async_playwright  
+from playwright.sync_api import sync_playwright  
 import re
 from openai import OpenAI
 from keys import OPENAI_API_KEY
@@ -44,31 +49,13 @@ def get_by_user_input(craft_data):
 
     return top_viewed
 
-async def auto_scroll(page, iterations=12, pause=0.75):
-    """Scroll to the bottom repeatedly to trigger lazy-loaded content."""
-    last_h = 0
-
-    for i in range(iterations):
-        await page.evaluate("window.scrollTo(0, document.body.scrollHeight)")
-        await asyncio.sleep(pause)
-        st.write(".", end="")  # Print dot for each scroll
-        h = await page.evaluate("document.body.scrollHeight")
-        if h == last_h:
-            st.write(" Done!")
-            break
-        last_h = h
-    else:
-        st.write(" Done!")  # If we completed all iterations
-
-async def scrape_URL_for_text(url):
-  
-    timeout = 60000  # milliseconds
-    st.write("Loading page content...")  # Start the progress line
-    async with async_playwright() as p:
-        # Run Playwright locally (no BrowserCat / no wss)
-        browser = await p.chromium.launch(headless=True)
-
-        context = await browser.new_context(
+def scrape_URL_for_text(url):
+    timeout = 60000
+    st.write("Loading page content...")
+    
+    with sync_playwright() as p:
+        browser = p.chromium.launch(headless=True)
+        context = browser.new_context(
             user_agent=(
                 "Mozilla/5.0 (Windows NT 10.0; Win64; x64) "
                 "AppleWebKit/537.36 (KHTML, like Gecko) "
@@ -76,15 +63,15 @@ async def scrape_URL_for_text(url):
             ),
             viewport={"width": 1366, "height": 900},
         )
-        page = await context.new_page()
-
+        page = context.new_page()
+        
         # Speed: skip heavy assets that aren't needed for text/HTML extraction
-        await context.route("**/*", lambda route: route.abort()
+        context.route("**/*", lambda route: route.abort()
                       if route.request.resource_type in {"image", "media", "font"}
                       else route.continue_())
 
         # Navigate and wait for the page to quiet down
-        await page.goto(url, wait_until="networkidle", timeout=timeout)
+        page.goto(url, wait_until="networkidle", timeout=timeout)
 
         # Try to dismiss any cookie banner (best-effort; ignore errors)
         for selector in [
@@ -93,29 +80,29 @@ async def scrape_URL_for_text(url):
             "text=/accept all cookies/i",
         ]:
             try:
-                await page.locator(selector).first.click(timeout=2000)
+                page.locator(selector).first.click(timeout=2000)
                 break
             except Exception:
                 pass
 
         # Trigger lazy-loaded content
-        await auto_scroll(page)
+        #await auto_scroll(page)
 
         # 1) Fully rendered HTML (serialized DOM).  <-- "true" HTML after JS runs
-        rendered_html = await page.content()
+        rendered_html = page.content()
 
         # 2) Readable text (tags stripped) - try to target the main article first
         selectors = ['article', 'main', '[role=main]', '[itemprop=articleBody]', '.article', '.post', '.entry-content', '.content']
         readable_text = ""
         for s in selectors:
             try:
-                el = await page.query_selector(s)
+                el = page.query_selector(s)
                 if el:
                     # Remove all anchor tags from this element before getting text
-                    await page.evaluate(f"""
+                    page.evaluate(f"""
                         document.querySelectorAll('{s} a').forEach(a => a.remove());
                     """)
-                    txt = (await el.inner_text() or "").strip()
+                    txt = (el.inner_text() or "").strip()
                     if len(txt) > 100:
                         readable_text = re.sub(r'\n{3,}', '\n\n', txt).strip()
                         break
@@ -124,7 +111,7 @@ async def scrape_URL_for_text(url):
 
         if not readable_text:
             # Fallback: clone body and remove likely non-article regions (header/footer/nav/aside/sidebar)
-            readable_text = await page.evaluate(
+            readable_text = page.evaluate(
                 """() => {
                     const clone = document.body.cloneNode(true);
                     const removeSel = ['header','footer','nav','aside','[class*="sidebar"]','[class*="cookie"]','[class*="masthead"]','a'];
@@ -145,7 +132,7 @@ async def scrape_URL_for_text(url):
         with open(fname, "w", encoding="utf-8") as f:
             f.write(readable_text)
 
-        await browser.close()
+        browser.close()
 
         return readable_text
 
@@ -212,12 +199,10 @@ def get_project(table):
     return project_title, url
 
 
-async def scrape_and_analyze(project_title, url):
-
-    
+def scrape_and_analyze(project_title, url):
     # Scrape the webpage for text content
     try:
-        text_content = await scrape_URL_for_text(url)
+        text_content = scrape_URL_for_text(url)
         
         if not text_content:
             return "ERROR: No text content could be scraped from the webpage"
@@ -235,11 +220,15 @@ def show_intructions(project_name, instructions_text):
     st.write(f"Instructions for {project_name}")
     st.text_area("Project Instructions", instructions_text, height=400, label_visibility="collapsed")
      
+# CH this ensures the table is locaded only once and cached
+@st.cache_data
+def load_data():
+    return pd.read_csv(r"data\projects_craft.csv")
 
 # MAIN
-st.title("Version 2 Progress Report A")
+st.title("Version 2 Progress Report A") # Please find a better title. What is your app called?
 
-craft_data = pd.read_csv(r"data\projects_craft.csv")
+craft_data = load_data()  # Only loads on first run, then cached
 
 if 'show_table' not in st.session_state:
     st.session_state.show_table = False
@@ -256,12 +245,8 @@ if st.session_state.show_table:
     if st.button("Get Instructions", type="secondary"):
         st.write(f"Selected project: {project_title}, please be patient while we fetch the instructions...")
         
-        # Use try-except to handle async issues
         try:
-            import nest_asyncio
-            nest_asyncio.apply()
-            instructions_text = asyncio.run(scrape_and_analyze(project_title, url))
+            instructions_text = scrape_and_analyze(project_title, url)  # No asyncio.run needed
             show_intructions(project_title, instructions_text)
         except Exception as e:
             st.error(f"Error fetching instructions: {str(e)}")
-            st.write("This might be due to async/Playwright compatibility issues with Streamlit.")
